refactor: extract demucs post-processing from worker
